chore: update gemini2.0-flash
The user wants to maximize the total value of a portfolio of publicly tradable assets.
I need to decide whether to BUY or SELL assets based on the current market conditions to maximize the portfolio value.
The portfolio currently holds AAPL.
I need to consider the current market conditions for AAPL and other assets to make a decision.
I should search for the current price of AAPL and other potentially better performing assets.
Since the net change of the operation should be zero, if I decide to buy another asset, I need to sell AAPL to get the funds.
I will check the current price of AAPL and compare it with other tech stocks like TSLA, NVDA, etc. to see if there are better investment opportunities.
I will use google search to get the latest stock prices.
After getting the prices, I will decide whether to keep AAPL or sell AAPL and buy another stock.
Let's search for current stock prices for AAPL, TSLA, and NVDA.